testing/web-platform/mozilla/tests/service-workers/service-worker/postmessage-to-client.https.html

<!DOCTYPE html>
<title>Service Worker: postMessage to Client</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>
var t = async_test('postMessage from ServiceWorker to Client');
t.step(function() {
    var scope = 'resources/blank.html'
    service_worker_unregister_and_register(
        t, 'resources/postmessage-to-client-worker.js', scope)
      .then(function(registration) {
          return wait_for_state(t, registration.installing, 'activated');
        })
      .then(function() { return with_iframe(scope); })
      .then(function(frame) {
          var w = frame.contentWindow;
          w.onmessage = t.step_func(onMessage);
          w.navigator.serviceWorker.controller.postMessage('ping');
        })
      .catch(unreached_rejection(t));

    var result = [];
    var expected = ['Sending message via clients'];

    function onMessage(e) {
      var message = e.data;
      if (message === 'quit') {
        assert_array_equals(result, expected,
                            'Worker should post back expected messages.');
        service_worker_unregister_and_done(t, scope);
      } else {
        result.push(message);
      }
    }
  });
</script>
